refactored pipeline code
