+++ cse/pipeline/Handler.py
@@ -1,0 +1,10 @@
+from cse.pipeline import Handler
+
+class SimpleConsolePrintHandler(Handler):
+    def process(self, ctx, data):
+        print(str(data))
+        ctx.write(data)
+
+class SimpleForwardHandler(Handler):
+    def process(self, ctx, data):
+        ctx.write(data)
